Module9/main (#17)
* update task requirement

* Separate core from tests

* remove unused package

* Creation of independent tests completed

* fix run issue

* fix test failure reasons

* fix
