+++ WebDriverTask.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 17.3.32929.385
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebDriverTask", "WebDriverTask\WebDriverTask.csproj", "{AA19A053-23F4-48A6-9967-2298B0244FA7}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GmailTest", "GmailTest\GmailTest.csproj", "{B2FE78D3-4D0A-4826-B784-63848A67F147}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{AA19A053-23F4-48A6-9967-2298B0244FA7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AA19A053-23F4-48A6-9967-2298B0244FA7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AA19A053-23F4-48A6-9967-2298B0244FA7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B2FE78D3-4D0A-4826-B784-63848A67F147}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B2FE78D3-4D0A-4826-B784-63848A67F147}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B2FE78D3-4D0A-4826-B784-63848A67F147}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B2FE78D3-4D0A-4826-B784-63848A67F147}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
