+++ WebDriverTask/Business/User.cs
@@ -1,0 +1,18 @@
+﻿namespace WebDriverTask.Business
+{
+    public partial class User
+    {
+        public string? FirstName { get; set; }
+        public string? LastName { get; set; }
+        public string? Email { get; set; }
+        public string? Password { get; set; }
+
+        public User(string? firstName = null, string? lastName = null, string? email = null, string? password = null)
+        {
+            FirstName = firstName;
+            LastName = lastName;
+            Email = email;
+            Password = password;
+        }
+    }
+}
