+++ WebDriverTask/Common/Pages/IPage.cs
@@ -1,0 +1,4 @@
+﻿namespace WebDriverTask.Common.Pages
+{
+    public interface IPage { }
+}
