--- WebDriverTask/Pages/Gmail/ContextMenu/Mail/MailContextMenu.cs
@@ -1,20 +1,0 @@
-﻿using OpenQA.Selenium;
-
-namespace WebDriverTask.Pages.Gmail.ContextMenu.Mail
-{
-    public class MailContextMenu : MailContextMenuElements
-    {
-        readonly IWebDriver webDriver;
-        public MailContextMenu(IWebDriver driver) : base(driver) => webDriver = driver;
-
-        public void ClickDelete()
-        {
-            DeleteItem.Click();
-        }
-
-        public void ClickArchive()
-        {
-            ArchiveItem.Click();
-        }
-    }
-}
