--- WebDriverTask/Pages/IPage.cs
@@ -1,4 +1,0 @@
-﻿namespace WebDriverTask.Pages
-{
-    public interface IPage { }
-}
